src/requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
cycler==0.11.0
easydict==1.9
Flask==2.1.2
fonttools==4.33.3
idna==3.3
imageio==2.19.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
kiwisolver==1.4.3
MarkupSafe==2.1.1
matplotlib==3.5.2
networkx==2.8.4
numpy==1.23.0
opencv-python==4.6.0.66
packaging==21.3
Pillow==9.2.0
PyMySQL==1.0.2
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
PyWavelets==1.3.0
requests==2.28.1
scikit-image==0.19.3
scikit-learn==1.1.1
scipy==1.8.1
six==1.16.0
threadpoolctl==3.1.0
tifffile==2022.5.4
torch==1.12.0
torchvision==0.13.0
tqdm==4.64.0
typing_extensions==4.3.0
urllib3==1.26.9
Werkzeug==2.1.2
zipp==3.8.0
